search: Move search code from message_view_header to search.js.
